CLI: Change default short option to display version from "-V" to "-v"
The current short option to display the version is "-V" (uppercase): `quarkus -V`. This is set by Picocli, but can be changed using the `picocli.version.name.0` property.

The problem with `-V` is that does not match with the other short options to display version in the extensions and other commands which is "-v".
Also, in other command tools like `mvn`, the option is `-v`.
I did check that the `-v` is not conflicting with other existing options.