Attempt to fix CT intermittent failures
Continuous testing tests can fail sometimes as an
additional unexpected run is triggered. This is because
a write is actually two seperate operations, a truncate
followed by a write, so the update processor can see this
as two seperate changes and trigger two seperate test
runs (depending on the timing). This fixes it by re-checking
the timestamps after compilation, and also attempting to
re-compile if they don't match.