OpenShift: Fix `quarkus.openshift.arguments` property
At the moment, when using the `quarkus.openshift.arguments` property and an unknown base image, it's replacing some arguments that are required for the Java command and hence making the application fail. For example, with no arguments are supplied, the OpenShift.yaml looks like:

```
command: java
args:
    - '-Dquarkus.http.host=0.0.0.0'
    - '-Djava.util.logging.manager=org.jboss.logmanager.LogManager'
    - '-jar'
    - /deployments/quarkus-run.jar
```

If we now set `quarkus.openshift.arguments=A,B`, then it changes to:

```
command: java
args:
    - 'A'
    - 'B'
```

This PR addresses the above situation by moving the required JAVA arguments to be part of the command and leaving the args field empty:

```
command: 
    - java
    - '-Dquarkus.http.host=0.0.0.0'
    - '-Djava.util.logging.manager=org.jboss.logmanager.LogManager'
    - '-jar'
    - /deployments/quarkus-run.jar
```

And now if we supply the arguments: `quarkus.openshift.arguments=A,B`, then:

```
command: 
    - java -Dquarkus.http.host=0.0.0.0 -Djava.util.logging.manager=org.jboss.logmanager.LogManager -jar /deployments/quarkus-run.jar
args:
    - 'A'
    - 'B'
```

Users would still be able to replace the whole command by using the command property: `quarkus.openshift.command` which I think it's the expected behaviour. Now the command and argument properties work the same in Kubernetes and OpenShift, regardless the base image we use for s2i. 

I've covered these changes using integration tests with the following scenarios:
- Kubernetes when using arguments
- Kubernetes when using command
- OpenShift when using arguments
- OpenShift when using commands

Also, I've verified these changes manually by deploying a sample application (Rest ++ Command Line application) and now it works fine.