Fix dev mode compile issues
If a file has a compile problem, and then another file is modified
that compiles correctly dev mode will restart and ignore the original
compile error.

This changes the compilation so that the timestamps are only updated
on a sucessful compile, so files that failed to compile will always
be recompiled.

(cherry picked from commit 0fcde158b7139cf3b037dbb1bee1994bcd91d585)