Update test with Mutiny for Hibernate Reactive
Injecting the session is discouraged because Quarkus will
behave differently when running the test from the command line
or when running it in the IDE.

In particular the tests will run in the wrong Vert.x context
when they are launched from the IDE, causing a failure.

This change makes it possible to run the test from the IDE without
errors.