Improve Kafka Avro serializer/deserializer autodetection
When the key/value class is `@AvroGenerated`, we autodetect
the corresponding Confluent / Apicurio Registry serde, which is
enough on the serialization side. On the deserialization side,
users had to add `specific.avro.reader=true` for Confluent or
`apicurio.registry.use-specific-avro-reader=true` for Apicurio
Registry. If they didn't, the deserializer would produce
a `GenericRecord`, which would always fail (we already detected
that users want the `@AvroGenerated` class).

With this commit, we automatically configure this property too,
and also support the `GenericRecord` class. This makes serde
autodetection for Avro more convenient.