Use testcontainers networking capabilities when using @QuarkusIntegrationTest + DevServices
This is done in order to get us off of host networking (which only works on Linux).
The way this works is that when @QuarkusIntegrationTest is meant to launch a container
and DevServices are needed, a build item is pushed into the build that forces
the DevServices to launch in TestContainers' SHARED Network.
The id of network is then used when launching the application as a container.

Fixes: #18194