Pass arguments arrays to Runtime::exec
Code scanning tools as in https://github.com/quarkusio/quarkus/security can flag this as a vulnerability.
In this case the URL is actually going to be safe, but it's still a better practice not to concatenate arguments as a String.

(cherry picked from commit 39ee796ce03148fdd953abe2d79ec24e2826611a)