Set the hang detection timeout to 1m for CI
None of our tests should really be doing nothing for one minute, if this
happens it is likely something is wrong.

Note that even if the timeout is triggered the test is not automatically
failed, a thread dump is just printed to make it possible to debug. Even
if a slow test causes this to be triggered it will not introduce
intermittent failures.