Make the build time defaults a higher priority
This means that if you set a value using a system property when building
the application it will now override the application.properties that is
packaged in the application (but not an external application.properties
on the file system).

Otherwise the behaviour is very confusing, with build time config values
only taking effect if they are not overriding a property in
application.properties.

Relates to (but does not fully fix) #19438