Fix Fault Tolerance in combination with RestClient Reactive
The fault tolerance extension loads bean classes from the deployment
classloader during fault tolerance discovery. The classes, as well as
discovered fault tolerance annotations, are transferred to runtime,
where SmallRye Fault Tolerance applies runtime configuration.

This works fine, unless one of the bean classes is generated
by another Quarkus extension. These generated classes are stored
in memory and the deployment classloader doesn't know about them.
Naturally, loading the class fails and aborts the build.

The specific problem found was in combination of Fault Tolerance
and RestClient Reactive. This commit makes sure that the RestClient
generated classes are known to the deployment classloader, but
it seems fairly obvious that this is a hacky one-off fix and
not a proper solution. I still find it worth sharing :-)