Don't interfere with testcontainers' ability to determine running containers
This is only applicable for containers running on a shared network.
Essentially we now ensure that for containers whose status can only
be determined by making a network call, that testcontainers can make
that network call.
In testcontainers 1.15.x, the library would actually not fail
if the connection could not be made, but would essentially give
up after a while.
With this change, startup of the tests is actually faster since testcontainers
actually determines the running status much faster.
In 1.16 the behavior seems to have changed and the testcontainers seems to try
to launch new containers and never gives up.