Change dev mode working directory
This will now be the same as surefire, so tests will see the same
user.dir for both surefire tests and continuous tests.

This is technically a breaking change, but should not really affect the
majority of applications. The old behaviour can be restored using

<workingDir>${project.build.directory}</workingDir>

in the maven plugin configuration.