CI - Always include a file at the root to get the full hierarchy
The upload-artifact action takes the common ancestor as the root path of
the archive which will lead to issues when we have only one file.

Always uploading a file from the root of the repository should fix this.

(cherry picked from commit 370d94426dd560618465f4682c05f2c77a12b889)