Fix invalid assertions in vert.x route tests
We were asserting that the error page contains the string "null", in
order to check that an NPE is reported.

But the error page contains javascript code, and that code contains
the javascript keyword "null".

So those assertions were always passing...
In some cases they were passing even though there was no NPE; the
assertion was incorrect and we didn't notice that because it was passing
anyway.