Add support for Dev Services inside Neo4j module.
This change adds one additional processor that conditionally (when docker is working and Neo4j is not reachable on the default address) starts up a Neo4j Test container based on the latest image.

Tests have been added that will be executed when docker is enabled.

The existing functional integration tests won't be affected as they use explicit configuration.