Improve order of checks and log level for Neo4j dev services
This changes the order of the checks for Neo4j dev services: First all explicit changes that have an effect on wether dev services are started or not are evaluated and if they lead to a non start, logged on debug.

If dev services should start (no changes or explicit config) but can’t as the default would connect to a locally running database as it used to be, a warning is logged.

Also: All tests are executed now.