Add a JTA aware variant of the Neo4j driver
This driver is aware of ongoing JTA transaction. If it detects an ongoing transaction, a session will be created accordingly that particpates in the tx.

To disable this in a bigger scenario (i.e. when an outer transactional method is calling a repo method that should independently commit), the standard means of `@Transactional(NOT_SUPPORTED)` (or `NEVER`) can be applied.