Fix race conditions in OIDC cache
The operator update based approach does not provide any real guarentees,
as multiple threads can be incrementing and decrementing at the same
time. compareAndSet must be used to guarentee sizing.

In addition this removes the attempt to invalidate an entry if the cache
is full. If the cache fills up then this will result in every request
iterating over every entry, which has pretty horrible performance
characteristics.