Fix setting class reader options
When multiple class transformations are registered, the class reader
options used to be overwritten by each transformation registration.
The last class transformation registered would control the class
reader options for all transformations of the same class. This caused
an issue for Logging with Panache, which requires the `EXPAND_FRAMES`
option.

This commit fixes that by merging the class reader options values,
so that all options that at least one transformation requires are set.
Since the class reader options are bit flags, merging them is a simple
bitwise OR.