Decouple rest-data-panache from RESTEasy Classic
This means that the `quarkus-hibernate-orm-panache`,
`quarkus-mongodb-rest-data-panache` and `quarkus-spring-data-rest`
no longer have hard dependencies on `quarkus-resteasy` and can thus be used with either `quarkus-resteasy` or `quarkus-resteasy-reactive`.

To work with RESTEasy Classic, the application much add the
`org.jboss.resteasy:resteasy-links` dependency (in addition to any `quarkus-resteasy*` dependency).
To work with RESTEasy Reactive, the application much add the
 `quarkus-resteasy-reactive-links` dependency.