Ensure the MockTracer is registered for tests
We need to do this, because the actual test runs in a separate ClassLoader
so there is no way of guaranteeing that some other Tracers will
not be registered before.

This worked before the change in Arc that stopped executing `<clinit>`
of beans only because the tests themselves are beans and therefore Arc
was forcing the `<clinit>`` of the test to be executed super early.