Add delay to ensure queue is created and bound.
In RabbitMQ queues are what store messsages for processing, while exchanges are what messages are sent to. If an exchange has no bound queues, in most standard cases, any messages sent to the exchange are dropped.

The "incoming" RabbitMQ connector is what creates and binds a queue to an exchange. This fix adds a delay to the "outgoing" messages to ensure the "incoming" connector has time to create and bind the required queue to the required exchange.