Add new BOOTSTRAP recording phase
This phase is guarenteed to run first when the application is started.
For JVM mode applications this means running before static init, for
native applications this means running before runtime init.

This allows logging to be setup as a priority, which means that logging
will work as expected while static init is being run. Without this
change applications that log a lot during static init (e.g. applications
with lots of Hibernate entities) will easily hit the dalyed handlers
limit, log messages will be dropped, and a warning printed to the
console.