Fix resolution of media type for Hibernate Validator in Resteasy Reactive
As the Resteasy Reactive is adding the `text/plain` media type for endpoints returning String, we can't infer the media type that is defined in the produces annotation. Therefore, we'll always supply the media type from the user accept header if it's supported, or JSON by default (as it was working before).