Fix CORS Filter to allow non-CORS OPTIONS requests
The previous implementation assumed that all OPTIONS requests were cors requests, and terminated them early, this fixes that, by ensuring that we're actually receiving a CORS request (presence of Origin and Access-Control-Request-* headers).

Included is a full set of test cases for the existing and changed behaviour. If you revert the change to CORSFilter.java, then you can see the bug.