REST Client config - read client configuration directly from Config
...rather than via `@ConfigRoot`

This should help in situations when REST Client config is obtained from
a vault ConfigSource, which doesn't list it's contents when
`ConfigSource#getPropertyNames()` is called. This causes the
configuration to not be included in the `@ConfigRoot` instance.

(cherry picked from commit 03ab3369a9756e920941dc21dd7e86f00e8d278a)