Reuse ClassPathUtils for finding flyway migrations The classpathutils already contains a nice implementation for searching a path in the classloader resources. It also handles local and jar files. We can just delegate the file search to it.
In the old implementation, a FileSystem was opened, but with "create:true", potentially creating new jar Files. This is not needed here, the jars with the migrations either already exist (prod build), or the migrations are present in directories (dev mode).

(cherry picked from commit 69f26bcd757fece47bf3ad81237bced8d17ef156)