AWT TS: Exceptions worded differently JDK11/17 vs. JDK18/19
Fixes #22067

An intentionally incorrect image is supposed to trigger an exception
about color profiles / palettes. The exception message states:

JDK 11, 17: "Can not access specified profile."
JDK 18, 19: "LCMS error 13: LUT is not suitable to be saved as LutBToA"

This commit improves the regular expression that checks the message.
It also replaces `|' with `█' as a separator in test data and utils
so as `|' doesn't break regular expressions parsing.