CDI context propagation improvements for the reactive stack
== ArC
Introduce the VertxCurrentContextFactory so that the Vert.x duplicated
context can be used to store the "current context" for normal scopes.

== resteasy-reactive
Don't clear our CDI current request when suspending.

This used to be done in order to prevent subsequent
requests running on the same thread as a suspended
request from accessing the former's data.
With the advent of DuplicatedContext backed
storage, there is no longer any chance of mixing
data so there is no need to clear it out.

Furthermore, by not clearing out current request,
code that accesses the request scoped CurrentVertxRequest
that is executed while the request is suspended, can now
work even if context propagation is not in play.

== Tests
Add leak detection tests for resteasy reactive, graphql and reactive rest client. Also improve the OpenTelemetry reactive tests.

== OpenTelemetry
Only register the OpenTelemetryClientFilter for RESTEasy Client Classic. Use Capabilities to determine when to register the OpenTelemetryClientFilter.

Co-authored-by: Georgios Andrianakis <geoand@gmail.com>
Co-authored-by: Clement Escoffier <clement.escoffier@gmail.com>
Co-authored-by: Roberto Cortez <radcortez@yahoo.com>
Co-authored-by: brunobat <brunobat@gmail.com>
Co-authored-by: Martin Kouba <mkouba@redhat.com>