Clean up how entries are ignored in jar building
The module-info.class ignore predicate was incorrect (it was testing if
"module-info.class" was ending with the path).
Also we used the list of ignored entries built to avoid duplicates in
uber jars for thin jar building. This was wrong as it could lead to us
ignoring the license for instance. Thin jar building shouldn't take this
list into account.