Unify instructions for both Mandrel and Oracle GraalVM
Mandrel has now matured and provides builds for both Linux and Windows
systems. Unifying the instructions makes the guide easier to follow and
removes the implication that Mandrel cannot be used instead of Oracle
GraalVM in some cases.

This commit also removes the hint for using Java 11 (instead of Java 8).
Java 8 is no longer supported and the community should be encouraged to
start testing Java 17.