Fix logging with panache in entities
In case an entity class uses `io.quarkus.logging.Log`, it requires
two bytecode transformations: the Hibernate transformation, and
the logging transformation. Unfortunately, `HibernateEntityEnhancer`
needs to break the chain of class visitors to be able to call into
the Hibernate Enhancer API. To be able to pretend that it's still
a regular `ClassVisitor`, it serializes the intermediate state
to a `byte[]`, runs the Hibernate enhancer, and then builds a whole
new `ClassReader` to continue in the original chain. This new
`ClassReader` is missing the options set on the original `ClassReader`,
which causes a failure in the logging transformation, because that
uses `LocalVariablesSorter`.

This commit introduces `QuarkusClassVisitor`, a new superclass
for class visitors that require access to some contextual data.
The bytecode transformation machinery puts the `ClassReader`
options into each `QuarkusClassVisitor` instance before
the bytecode transformation process begins. Finally,
`HibernateEntityEnhancer` extends this new class, so that when
it creates the new `ClassReader`, it can pass the correct options.

(cherry picked from commit b87454d20901142d1fa19b364fced0fed4e06731)