Copy RestClient interface annotations to generated class
When RestClient Reactive generates the implementation class, it does
not copy interface-level annotations to the generated class (though
it does copy annotations from RestClient methods). This means that
interceptor bindings declared on the interface itself don't work.

With this commit, RestClient Reactive will copy all annotations
declared on the RestClient interface (with the exception of
a scope annotation and some MicroProfile RestClient annotations)
to the generated class. This includes interceptor bindings, too.