fix cli install
Details in https://github.com/quarkusio/quarkusio.github.io/pull/1307
on why the "trust" call is added.

I moved jbang up as it is the only one that works on every OS and without
java first installed. Thus absolute fastest/simplest to use on all major OS's.

And moved homebrew down as that is only available for mac/osx.