Resteasy Reactive: Aggregate part items when using @MultipartForm
When using standard HTML select[multiple]:
```html
<form ...>
    <select name="values" multiple ...>
    </select>
</form>
```

The request that HTML is doing to the server looks like: `values=X&values=Y&...` (if X and Y was both selected).

The problem happens in the server side, when using resteasy reactive:

```
@POST
    @Consumes(MediaType.MULTIPART_FORM_DATA)
    @Path("/endpoint")
    public String greeting(@MultipartForm FormData formData) {
        // formData.values fails!
        // ...
    }
```

And FormData:

```
public class FormData {

    @RestParam
    @PartType(MediaType.TEXT_PLAIN)
    public String[] values;
}
```

This fails with a serialization error (it cannot find a serializer that works with arrays). (Related stackoverflow question: https://stackoverflow.com/questions/67151539/multiple-file-upload-using-quarkus-and-resteasy - for Resteasy classic).

However, we could make it work with arrays to support the above scenario with the changes part of this PR.

Note that I did check that it works fine with complex objects like Person.