Inject "annotations" and "genericType" in ParamConverterProvider
Note that this PR will introduce a performance penalty at load time (first time, when the rest client instance is loaded AND if and only if the rest client methods use param annotations - which is most of the times tho).

What I did was to always generate the `javaMethodX` static fields (before, it was also generated always, but it was done in the implementation MicroProfileRestClientEnricher).
Plus, apart of the method information, we will also load the annotations and the generic types. This is done at load class time (static init).

Moreover, as I had to move some code from MicroProfileRestClientEnricher to
JaxrsClientReactiveProcessor, in order to not increase the length of this class JaxrsClientReactiveProcessor, I started moving some code to ClassRestClientContext (which is protected - not available for users)

Fix https://github.com/quarkusio/quarkus/issues/22870