fix(quartz): fix NPE when running quartz in database mode and migrating from old versions
Fixes https://github.com/quarkusio/quarkus/issues/23457.

When migrating from versions which are < 2.1 to 2.7, a NPE was produced
due to trigger name renaming to remove the `_trigger` suffix.

The fix in here introduce a migration support to reschedule old triggers.
I went this way to reduce the toil that the user will have to do performing the actions described in our migration documentation.
This is only an exception case where the extension will have to perform the migration.
The temporary code will reschedule the old trigger i.e the one with the _trigger suffix to a new one without the suffix.
This shouldn't impact in any way the application and should be transparent to the user.

I've tested the fix against the reproducer given
https://github.com/quarkusio/quarkus/issues/23457#issuecomment-1035483218
and the verification seemed conclusive to me.