Update working directory for Jib created images
With the use of ubi8/openjdk base images as the default,
we now also update the working directory to avoid
any permissions issues (as the containers are run by a
non-root image).

This change is compatible with the pre Quarkus 2.7
base images as well (for those users who wish to continue
using that base image).