Fix setting of JAEGER_SAMPLER_MANAGER_HOST_PORT system property
Originally, the property had been set to the value returned by
InetAddress.toString(), which always includes the resolved IP address
(or the string "unresolved") in addition to the original host name.
For example, setting "quarkus.jaeger.sampler-manager-host-port" to
"my-jaeger-host:5778" would result in a system property value of
"my-jaeger-host/<unresolved>:5778" which cannot be used by the Jaeger
client to retrieve the sampler strategies.

The property isno set to the original host string + ":" + port instead.

Fixes #23711

Signed-off-by: Kai Hudalla <sophokles.kh@gmail.com>