Ensure that remote-dev works with containers created by Jib
The issue without this is that Jib overwrites the owner
of the working directory and sets it to root.
With this fix, we add a faux layer that simply
reverts the owner back to the default when the
default base image is used.

Originally discussed at: https://quarkusio.zulipchat.com/#narrow/stream/187038-dev/topic/remote.20dev.20in.202.2E7/near/273449615