OIDC: Allow authenticating AccessTokenCredential without routing context
The `quarkus-oidc` extension should allow the user to perform a programmatic authentication using the `IdentityProviderManager#authenticate()` and `IdentityProviderManager#authenticateBlocking()` APIs. One use case for this is when the application temporarily stores a given token and uses it later (possibly in a different thread outside of the scope of the original request) to perform some action.