Document| regex capabilities for CORS origins
The current documentation on the CORS filter at https://quarkus.io/guides/http-reference#cors-filter only instructs about the regular expression capabilities for origins in the example, but not in the reference section for available config properties. This commit enhances the javadoc comment to include a hint about the regular expression support.