Support multiple levels of JUnit Nested with QuarkusTest
The initTestState is being called each time there is a Nested class, so we need to keep the previous reference in the outerInstances list. 
Only the last class (the one that has no more enclosing classes) is the actual test instance and the one that can be resolved using CDI (runningQuarkusApplication.instance(outerClass)).

Fix https://github.com/quarkusio/quarkus/issues/19845