Oracle Dev Service - Fix shared network JDBC URL
Use a JDBC URL that is consistent with the one used when we don't use a
shared network (i.e. exactly the same as generated by the container
getJdbcUrl() method except for the host and port).
At the moment, it is using a SID URL, which is inconsistent and looks
incorrect to me.