Allow passing a string to QuarkusApplicationClassBuildItem
This is especially useful when you generate the @QuarkusMain class.
The @QuarkusMain annotations are looked for in the combined index,
thus not taking into account generated classes.
The only way to push a generated class as @QuarkusMain is to use
QuarkusApplicationClassBuildItem and obviously, in this case, you don't
have a Class handy.