Make sure Kafka serde autodetection doesn't override connector configuration
The SmallRye Reactive Messaging Kafka extension automatically detects
key and value serializer/deserializer and sets default config values
for the respective channel. That overrides connector configuration,
which is unexpected. If the user provides connector configuration
themselves, we should not override that. Unfortunately the extension
can only detect connector configuration if it's present during build,
but that will be the case 99.9% of time.

With this commit, if connector configuration is present for key/value
serde, the respective channel configuration is not emitted.

(cherry picked from commit 366ea7430df9939292cd0ac7f4dae3df7c503a93)