Add REST Client Reactive extensions for JAXB and JSON-B
I have added one simple test and no integration tests for now as things
are covered by the server tests given how things work and I wasn't sure
it was worth adding 10 minutes to the build just for that.
But I could be convinced otherwise if someone has a strong opinion about
it.