Pass pre-generated proxies to HibernateOrmRecorder as a constant
This is a hack to avoid introducing circular dependencies between build steps.

If we just passed the proxy definitions to #build as a normal build item,
after the changes in the next commits,
we would end up with the following dependencies:

  #pregenProxies => ProxyDefinitionsBuildItem => #build => BeanContainerListenerBuildItem
  => Arc container init => BeanContainerBuildItem
  => some RestEasy Reactive Method => BytecodeTransformerBuildItem
  => build step that transforms bytecode => TransformedClassesBuildItem
  => #pregenProxies

Since the dependency from #preGenProxies to #build is only a static init thing
(#build needs to pass the proxy definitions to the recorder),
we get rid of the circular dependency by defining a constant
to pass the proxy definitions to the recorder.
That way, the dependency is only between #pregenProxies
and the build step that generates the bytecode of bytecode recorders.

See https://quarkusio.zulipchat.com/#narrow/stream/187038-dev/topic/Recorder.20introducing.20cyclic.20dependency.20between.20build.20steps