Use container runtime detection in @QuarkusIntegrationTest
This is done with the same code that is used during build time.
For this to work, the code that perform the detection was moved
from the build time module to the runtime module.

(cherry picked from commit f43797f2a367e19c3ed9dd70ada78fb35c5d9c5a)