Add QuarkusClassLoader#isResourcePresentAtRuntime()
This method is used to check that a resource (class included) is
available at runtime.
It is especially important to use this when checking if a class is
available in a build step because using the TCCL to do that will lead to
checking the deployment class loader which also contains all the
deployment dependencies - dependencies which might not be available at
runtime.