Kubernetes: Merge StatefulSet resources from input if set
When users add a StatefulSet resource and the deployment-target is set to `StatefulSet`, we need to read the StatefulSet resource from the user and populate it accordingly.

Before these changes, we were including the StatefulSet resource, but adding a new one (with the same name), so in the end, two StatefulSet resources were included in the final kubernetes.yml.

Fix https://github.com/quarkusio/quarkus/issues/25162