AppCDS: Make container-based generation opt-out.
This change adds a new configuration key,
quarkus.package.appcds-use-container, which defaults to true and can be
set to false to avoid the use of a container when building an AppCDS
archive.

This is useful in cases where the build happens in an environment
without full control of a container system, such as hosted CI solutions.