Fix FileTestCase.testFiles on windows with core.autocrlf input File size is dependent on how the project was checked out. When git is set up with core.autocrlf input, the checkout keeps the line endings as they are commited to the repository (as LF).
On windows, the default System line seperator however is CTRLF.
Previosly, CTRLF was used to replace LF, since the assumption was that the checkout matches the System line seperator.
This is not true however if anything other than the default was choosen for core.autocrlf. On windows, this is quite easy to achieve, since git for windows even has a dialog in the installation wizard for this setting.

This commit now changes the test to simply always read the file from the FS, instead of keeping a copy of the files content as a static variable. This solution is now agnostic to line seperators.